Health-prediction-of-fetus-using-ctg-main/requirements.txt
numpy
pandas
scikit-learn
tensorflow
joblib
streamlit
matplotlib
setuptools
plotly
seaborn
reportlab
cryptography
pyyaml